优化接口和方法，完备检查所有表的关联性问题

+++ abnormalityAccept/src/main/java/com/abnormality/abnormalityaccept/service/QuestService.java
@@ -17,6 +17,6 @@ import com.github.pagehelper.PageInfo;
      boolean updateQuest(Quest quest);
 
      boolean deleteQuestById(Long id);
-     boolean addQuest(Quest quest, Integer sendId);
+     boolean addQuest(Quest quest, Long sendId);
      PageInfo<Quest> findQuestByConditions(QuestParam questParam);
 }
